Fix [Clients]: paging thing that was broken

+++ app/src/main/java/com/poli/prevendasomie/navigation/Screen.kt
@@ -21,7 +21,7 @@ sealed class Screen(val route: String) {
 
     object ClientDetailScreen : Screen("$CLIENT_DETAIL_ROUTE/{clientId}") {
 
-        fun passClientId(clientId: Int): String {
+        fun passClientId(clientId: Long): String {
             return "$CLIENT_DETAIL_ROUTE/$clientId"
         }
     }
